refactor: Changed server to next.js api route
