+++ components/analyzer.tsx
@@ -29,9 +29,7 @@ export default function Analyzer({
     onLoading(true);
     try {
       const response = await fetch(
-        `http://127.0.0.1:5001/words?text=${encodeURIComponent(
-          formData.textToSplit
-        )}`,
+        `/api/words?text=${encodeURIComponent(formData.textToSplit)}`,
         {
           method: "GET",
         }
