+++ package.json
@@ -20,6 +20,7 @@
     "csv-writer": "^1.6.0",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
+    "express-validator": "^7.2.0",
     "lucide-react": "^0.414.0",
     "mecab-async": "^0.1.2",
     "next": "14.2.5",
@@ -29,6 +30,7 @@
     "react-hook-form": "^7.52.1",
     "tailwind-merge": "^2.4.0",
     "tailwindcss-animate": "^1.0.7",
+    "validator": "^13.12.0",
     "xml2js": "^0.6.2",
     "zod": "^3.23.8"
   },
